+++ assignments/solutions/files-match-result/examples/step_1.rs
@@ -1,10 +1,10 @@
 use std::fs::File;
 
 fn main() {
-    let open_result = File::open("src/lib/content.txt");
+    let open_result = File::open("src/data/content.txt");
 
     match open_result {
         Ok(_file) => println!("File opened"),
-        Err(e) => println!("Problem opening the file: {:?}", e),
+        Err(e) => panic!("Problem opening the file: {:?}", e),
     };
 }
