+++ assignments/solutions/files-match-result/examples/step_3.rs
@@ -1,25 +1,21 @@
 use std::fs::File;
-use std::io::prelude::*;
-use std::io::Error;
-
-fn unwrap_file(open_result: Result<File, Error>) -> File {
-    match open_result {
-        Ok(file) => return file,
-        Err(e) => panic!("Problem opening the file: {:?}", e),
-    };
-}
-
-fn content_to_string(mut file: File) -> Result<String, Error> {
-    let mut content_string = String::new();
-    file.read_to_string(&mut content_string)?;
-    Ok(content_string)
-}
+use std::io::{BufRead, BufReader};
 
 fn main() {
-    let open_result = File::open("src/data/content.txt");
+    let open_result = File::open("./src/data/content.txt");
 
-    let file = unwrap_file(open_result);
+    let file = match open_result {
+        Ok(file) => file,
+        Err(e) => panic!("Problem opening the file: {:?}", e),
+    };
 
-    let content = content_to_string(file).unwrap();
-    println!("{}", content);
+    let buf_reader = BufReader::new(file);
+
+    let mut number = 0;
+
+    for _line in buf_reader.lines() {
+        number += 1;
+    }
+
+    println!("{}", number);
 }
