+++ assignments/solutions/files-match-result/examples/step_4.rs
@@ -1,25 +1,24 @@
 use std::fs::File;
-use std::io::{BufRead, BufReader, Error};
-
-fn unwrap_file(open_result: Result<File, Error>) -> File {
-    match open_result {
-        Ok(file) => return file,
-        Err(e) => panic!("Problem opening the file: {:?}", e),
-    };
-}
+use std::io::{BufRead, BufReader};
 
 fn main() {
     let open_result = File::open("src/data/content.txt");
 
-    let file = unwrap_file(open_result);
+    let file = match open_result {
+        Ok(file) => file,
+        Err(e) => panic!("Problem opening the file: {:?}", e),
+    };
 
     let buf_reader = BufReader::new(file);
 
-    let mut number = 0;
-
-    for _line in buf_reader.lines() {
-        number += 1;
+    for line in buf_reader.lines() {
+        match line {
+            Ok(content) => {
+                if !content.is_empty() {
+                    println!("{}", content)
+                }
+            }
+            Err(e) => println!("Error reading line {}", e),
+        }
     }
-
-    println!("{}", number);
 }
