+++ assignments/solutions/files-match-result/examples/step_5.rs
@@ -1,28 +1,46 @@
 use std::fs::File;
-use std::io::{BufRead, BufReader, Error};
+use std::io::{BufRead, BufReader};
+use url::Url;
 
-fn unwrap_file(open_result: Result<File, Error>) -> File {
-    match open_result {
-        Ok(file) => return file,
-        Err(e) => panic!("Problem opening the file: {:?}", e),
-    };
+fn parse_line(line: String) -> Option<Url> {
+    match Url::parse(&line) {
+        Ok(u) => Some(u),
+        Err(_e) => None,
+    }
 }
 
 fn main() {
     let open_result = File::open("src/data/content.txt");
 
-    let file = unwrap_file(open_result);
+    let file = match open_result {
+        Ok(file) => file,
+        Err(e) => panic!("Problem opening the file: {:?}", e),
+    };
 
     let buf_reader = BufReader::new(file);
 
     for line in buf_reader.lines() {
-        match line {
-            Ok(content) => {
-                if !content.is_empty() {
-                    println!("{}", content)
-                }
-            }
-            Err(e) => println!("Error reading line {}", e),
+        let line = match line {
+            Ok(content) => content,
+
+            Err(e) => panic!("Error reading line {}", e),
+        };
+
+        let url = parse_line(line);
+
+        match url {
+            Some(line) => println!("{}", line),
+            None => continue,
         }
     }
+}
+
+#[test]
+fn correct_url() {
+    assert!(parse_line(String::from("https://example.com")).is_some())
+}
+
+#[test]
+fn no_url() {
+    assert!(parse_line(String::from("abcdf")).is_none())
 }
